feat: add mapper.findMany iterator and fix UnixMillis schema mapping
- add mapper.Rows(T) iterator with next()/deinit() backed by prepared Stmt
- implement mapper.findMany(T, P, db, allocator, where_clause, params) using T.Meta table and Stmt.bindAll
- add mapper.freeOwned(T, allocator, *T) to free owned slice(u8) fields produced by row decoding
- fix schema type mapping for types.UnixMillis to generate INTEGER columns
- add integration test for findMany iteration with ORDER BY filter and owned field freeing
- update itest aggregator to include findMany integration test

+++ src/wrapper/stmt.zig
@@ -72,7 +72,7 @@ pub const Stmt = struct {
             return error.SqliteBindFailed;
     }
 
-    pub fn bindFloat(self: *Self, idx: c_int, value: i64) !void {
+    pub fn bindFloat(self: *Self, idx: c_int, value: f64) !void {
         const rc = raw.sqlite3_bind_double(self.stmt, idx, @as(f64, @floatCast(value)));
         if (rc != db_ok)
             return error.SqliteBindFailed;
